Add scripts for manual multiple task execution on HTCondor
HTCondor allows submitting large amounts of jobs but it's very slow when
doing that. The multi-rail-estimate scripts instead take a different
approach, submitting a smaller amount of tasks but executing the main
program multiple times per task. This way it's possible to keep the
slowness of the HTCondor scheduler bounded.