Add scripts for manual multiple task execution on slurm
slurm doesn't allow submitting large amounts of jobs because it works
with requests for a specific number of computing nodes, not tasks (and
the number of nodes might be orders of magnitude smaller than the number
of tasks). It supports submitting task arrays, which provide equivalent
functionality, but task arrays have the side-effect that it adds to the
scheduler task backlog, which is typically restricted due to quota
limits of the cluster provider.

The multi-hello scripts are an example of working around the limitation
by manually executing the main program multiple times per task. This
way, it's possible to request a small number of nodes and execute a
large number of tasks.