Add script to calculate total execution time of scripts run on cluster
This script uses preformatted lines in log files to extract the start
and finish times, calculating each duration and summing all durations to
produce a total execution time across the whole cluster.

The script is designed to not depend on a specific scheduler and avoid
considering dead time caused by the scheduler, so other scripts must
adhere to the preformatted line specification dictated by this script.
The format is the following:

    <date>: <event>: <description>

The precise formats that can be parsed are defined by a regular
expression in the code. For short, date must be in ISO standard format,
the event must be 'Starting' or 'Finished' and the description field
must be the same in both start and finish events.

The script only considers the first starting event in the file and its
corresponding finish event, ignoring all other events.

Execution syntax is the following:

    $ rail-sum-execution-time <file-pattern>...

The parameters may be file names or glob expressions. There must be one
file per task. Each file must contain a start and finish event line.