Limit memory and CPU usage on rail-condor
The RAIL framework uses many auxiliary libraries and some of them, in
particular numerical computing and file I/O libraries may spawn a large
number of auxiliary threads and processes. When running on the cluster,
this is undesirable because each estimation process runs on a single
cluster thread and auxiliary threads will not help speeding up I/O or
computation, while having the collateral effect of increasing memory
usage. This patch limits both memory and CPU usage to the running
processes to moderate values, so that memory is preserved. The patch
adds the following configurations:

- Limits maximum virtual memory with setrlimit() (currently 16 GiB)
- Limits the maximum CPUs used with sched_setaffinity (tries to limit
  usage to 2 Hyperthread siblings and 2 cores sharing the same cache)
- Sets the OMP_NUM_THREADS environment variable (currenly 3)
- Sets the ARROW_IO_THREADS environment variable (currently 3)
- Sets the default Arrow memory pool backend to "system" (malloc())

This patch fixes the memory explosion seen in large tests in the Apollo
cluster.