Increase the out-of-memory killer score of the rail-estimate process
This patch tries to make the rail-estimate processes running in the cluster more
suitable to be killed by the linux kernel when an out-of-memory condition
happens. Consequently, it's expected that the slurm daemon will become
relativelly less suitable to be killed and may continue working even on low
memory conditions.