Make rail-slurm retry a few times after failed executions
This patch makes the rail-slurm scripts more tolerant to cluster problems.
Whenever an execution of the srun or rail-estimate programs return failure, the
execution request is added back to the queue and tried again without canceling
the whole job. After a few unsuccessful retries, the script gives up as usual
and cancels the execution.

Currently, the implementation attempts 3 executions per file.