Implement simple I/O benchmark to help with debugging
debugstorage is a simple read or write performance measurement program,
that has low memory requirements and requires no dependencies. It writes
random data sequentially in a loop to a set of files, or reads them
back, also sequentially. Memory is allocated once for the data, with the
size given as a command-line parameter, and locked to disallow being
swapped to disk and then a few times to record timing data in a very
small event array. The program executes the benchmark, measuring the
time that each I/O operation takes and in the end prints them.

Usage:
  debugstorage [-r|-w] [-s <size>] [-l <loop>] [<fileprefix>]

Options:
 <fileprefix>	Text used when creating file names [default: "random"].
 -r		Read existing files.
 -w		Write new files [default].
 -s <size>	Set the file size [default: 134217728].
 -l <loop>	Repeat the I/O operation [default: 5].

File names will have the format: "<fileprefix>-<loopiter>.out".
All files will have the same random content.
All memory allocated for the requested size will be locked.

Compile with:

	$ cc -o debugstorage debugstorage.c

Example usage:

	$ debugstorage -w -s 65536 -l 10 testfiles
	$ debugstorage -r -s 65536 -l 10 testfiles