Implement program to protect the slurm daemon from the job process
slurm-shield - execute program on slurm cluster while protecting slurm from it */

Executes and monitors a child process with lower CPU and I/O priority and higher OOM
adjustment score relative to the slurm daemon, so that there's a decreased probability
of the slurm daemon being killed by the operating system and an increased probability
of the child program being killed in a low memory situation. The child process is also
monitored and checked for being terminated by signals and then hiding those abnormal
terminations from the slurm daemon by returning a simple error result. Similarly, it
monitors signals being received by itself, forwarding them to the child and following
them with a SIGKILL, as a best effort to terminate a stuck child process.

Usage:

	slurm-shield <program> [<args>...]

The program is searched using the PATH environment variable.

Compile with:

	$ cc -o slurm-shield slurm-shield.c

Example usage:

	$ slurm-shield rail-estimate -a fzboost src.pq dest.hdf5