Manually configure joblib auxiliary library
Newer versions of joblib allow globally configuring the execution mode.
This patch changes from the default multiprocess mode to multithreaded
to avoid a side-effect of the joblib implementation: it uses
memory-mapped I/O to communicate between processes, which is too heavy
for the rail-estimate use-case.

This was shown to reduce the number of total auxiliary processes and
threads and also increase speed.

The patch also sets the number of auxiliary threads to 1, but it's known
that some libraries overwrite this value, so it won't always work.