Use different default slurm parameters for TPZ
When running TPZ estimations with 100 trained decision trees, the TPZ
algorithm's memory requirements are 3500 MB per process. This is bigger
than the maximum 2140 MB limit available in the LineA cluster. This
patch changes the default TPZ configuration to ask for a smaller number
of harts per machine, so that more memory can be assigned to each
process. The new TPZ default parameters are the following:

    -N 26 -n 1316 --mem-per-cpu=3500M

The number of tasks is calculated as 14 old machines running 34 tasks
plus 12 new machines running 70 tasks (14×34 + 12×70 = 1.316).