Support passing through arbitrary parameters in pz-compute.batch
Previously, the pz-compute.batch slurm script only allowed passing the
estimation algorithm and optionally a parameter file name to
rail-estimate. This patch changes this to allow passing an arbitrary
number of parameters. The command-line given to the script is just
forwarded as is to rail-estimate.

Example usage:

    $ sbatch rail-slurm.batch input output -a bpz -g group1 -b 101

In the example, the algorithm, HDF5 group and number of bins are
forwarded to rail-estimate.