Store pz-compute's full command line into a log file
This patch makes pz-compute save a YAML file storing the program's argv
in it. Before saving, it also tries to read any old log file present to
compare the old command line to the new one. This patch will be useful
for further patches that try to resume some previously interrupted
execution of pz-compute: the comparison between the old and new command
lines might be used to determine if the new execution should be
considered a retry or not.