Skip already processed files when calling pz-compute (support resume)
This patch adds support for inspecting the output directory for output
files that have already been processed by some previous execution of
rail-estimate. Whenever some output file with a name corresponding to an
input file exists, it's considered that the input has already been
previously processed and its output is available, so that it won't be
processed again. This effectively adds support for resuming a previous
execution of pz-compute that might have been interrupted in the middle.

The patch relies on the fact that the output file must be correct and
complete. This is the case when executing the RAIL estimations, since
the default ceci stage uses a temporary file name while writing the file
and then (atomically) renames the file after the estimation is complete.

The patch also deletes temporary files encountered in the output
directory.