Disable slurm's automatic CPU affinity configuration
The pz-compute.run script already modified the CPU affinity to restrict the
number of harts used by rail-estimate, but in practice it was receiving a
previously configured CPU affinity mask from slurm. When running sbatch with
certain parameters, though, slurm is auto-configuring the CPU affinity mask in
unexpected ways (in particular, when using -c2, all processes are being
allocated to a single socket and the other socket on the LineA's machines are
being kept idle, even when all are requested).

This patch completely disables slurm automatic CPU affinity preconfiguration.
Custom configuration in pz-compute.run is kept the same.