Remove preset of OMP_NUM_THREADS environment variable
This variable was hardcoded to 2, but should be changed together with
the CPU affinity mask, configured by the pz-compute.run script (which
will become more complex). gcc's libgomp respects the CPU affinity value
for setting the default value of this variable, so just removing it will
work as intended on such cases.