Remove the default 16 hour execution limit from pz-compute
There are some algorithms that will take more than that in large
estimations. Without this setting, the maximum timeout allowed by the
slurm partition will be requested instead. A specific time limit may be
requested with sbatch_args and a new easy to use time limit parameter
will also be added in a later patch.