Config default LEPHAREDIR and LEPHAREWORK environment variables
LePhare depends on a correctly set lephare-data dataset repository and
on a correctly set cache directory. If the user does not provide a
pre-configured LEPHAREDIR environment variable, then set to the default
lephare-data location in the LineA storage. If the user does not provide
a LEPHAREWORK (or XDG_CACHE_HOME) environment variable and it's not
present in the YAML configuration file, then set the default value to
'lephare-work/train'.